# openapi.yaml
openapi: 3.1.0
info:
  title: OpenTerms Agent Action Observability API
  version: 0.2.0
  summary: Backend API for ingesting, evaluating, and querying ORS-conformant agent action receipts.
  description: |
    The OpenTerms Agent Action Observability API is the backend for the
    OpenTerms Agent Action Observability product. It accepts cryptographically
    signed receipts that conform to the Open Receipt Specification (ORS),
    evaluates them against a deterministic policy engine, persists them to an
    append-only log, and exposes query, simulation, and key-management endpoints
    for audit, compliance, and policy iteration.

    # ORS version target

    This contract accepts the **union of ORS v0.1 and ORS v0.2**. ORS v0.2 is
    fully backward compatible with v0.1 (the domain-separation prefix and signing
    algorithm are unchanged), so a single ingest pipeline validates both. v0.2
    optional payload fields (`terms_type`, `terms_service`, `terms_version`),
    the `model_training` action type, and the normative `ors.anchor` and
    `ors.zk_proof` extensions under `action_context.ors` are documented in this
    contract but are not required.

    Sources of truth:

    - ORS v0.1 spec: https://github.com/jstibal/ors-spec/blob/main/ORS-v0.1.md
    - ORS v0.2 spec: https://github.com/jstibal/ors-spec/blob/main/ORS-v0.2.md
    - Reference verifier (Python): https://github.com/jstibal/ors-spec/blob/main/verify.py
...

    # Receipt shape

...

...

...

    # Append-only semantics

...

...

...
servers:
...
security:
  - bearerAuth: []
  - oauth2: [read, write]

# ============================================================================
# Tags
# ============================================================================
...
paths:
...
  /receipts/ingest:
    post:
      tags: [Receipts]
      summary: Ingest a signed ORS receipt
      description: |
        Accepts a signed ORS receipt (v0.1 or v0.2). The service:

        1. Validates the receipt against the ORS payload schema.
        2. Recomputes the canonical hash per ORS Section 4 and compares it to
           the receipt's `canonical_hash`.
        3. Looks up the public key by `key_id` in the workspace JWKS.
        4. Verifies the Ed25519 signature over `b"ORSv0.1\x00" + hash_bytes`.
        5. Deduplicates by `canonical_hash` unique index.
        6. Persists the verified receipt to the append-only log.

        Receipts that fail any validation, hash, or signature check are NOT
        written to the receipts table. Failed verifications log to a separate
        error stream and return a 422.

        Use the `Idempotency-Key` header to make retries safe. A replay with
        the same key and the same canonical hash returns the stored receipt. A
        replay with the same key and a *different* payload returns 409 with
        `IDEMPOTENCY_KEY_CONFLICT`.

        Receipts are immutable once persisted. There is no endpoint to modify
        or delete a receipt.
      operationId: ingestReceipt
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignedReceipt'
            examples:
              basicApiCall:
                $ref: '#/components/examples/BasicApiCallReceipt'
              dataAccessWithCommitments:
                $ref: '#/components/examples/DataAccessWithCommitmentsReceipt'
              policyClassificationV02:
                $ref: '#/components/examples/PolicyClassificationReceipt'
      responses:
        '201':
          description: Receipt ingested and persisted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '200':
          description: |
            Receipt was already ingested (duplicate by canonical hash). Returns
            the stored receipt unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Idempotency key collision with a different payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                idempotencyConflict:
                  value:
                    error:
                      code: IDEMPOTENCY_KEY_CONFLICT
                      message: An earlier request with this Idempotency-Key was processed with a different payload.
        '422':
          description: Receipt failed ORS schema validation, hash recomputation, or signature verification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                hashMismatch:
                  value:
                    error:
                      code: HASH_MISMATCH
                      message: Recomputed canonical hash does not match the value in the receipt.
                      details:
                        expected: d486a0a0d298a91b52544aea298693f0ea3584d5f808955d7c09d5f99c16ee32
                        computed: 8b1c4e2a91f0...
                signatureInvalid:
                  value:
                    error:
                      code: SIGNATURE_INVALID
                      message: Ed25519 signature failed to verify against the workspace JWKS public key.
                      details:
                        key_id: key_abc123def456
                unknownIssuer:
                  value:
                    error:
                      code: UNKNOWN_ISSUER
                      message: The key_id on the receipt does not match any key in the workspace JWKS.
                      details:
                        key_id: key_2024_09_15_b1c2
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /receipts/verify/{hash}:
    get:
      tags: [Receipts, Public]
      summary: Verify a receipt signature (public)
      description: |
        Public endpoint, CORS-open, no authentication required. Recomputes the
        canonical hash on the stored receipt and verifies the Ed25519 signature
        against the issuer's JWKS public key. Returns a verification result
        whose shape mirrors the reference Python verifier
        (`github.com/jstibal/ors-spec/blob/main/verify.py`).

        Path follows the ORS spec convention (`/v1/receipts/verify/{hash}`) for
        ecosystem interop with provider verification flows described in ORS
        Section 9. The earlier draft path `/v1/receipts/{hash}/verify` is not
        served.

        This endpoint exists so any third party (auditor, regulator, API
        provider, counterparty) can confirm receipt authenticity without an
        account.
      operationId: verifyReceipt
      security: []
      parameters:
        - $ref: '#/components/parameters/ReceiptHash'
      responses:
        '200':
          description: Verification result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationResult'
        '404':
          $ref: '#/components/responses/NotFound'

  # --------------------------------------------------------------------------
  # Decisions
  # --------------------------------------------------------------------------
...
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: |
        Workspace-scoped API key. Used by SDK consumers (`openterms-py`,
        `openterms-ts`, LangChain and CrewAI adapters). Pass as
        `Authorization: Bearer <api_key>`.

    oauth2:
      type: oauth2
      description: |
        OAuth 2.0 authorization-code flow for dashboard and interactive admin
        users. Issued by the OpenTerms identity provider.
      flows:
        authorizationCode:
          authorizationUrl: https://observe.openterms.com/oauth/authorize
          tokenUrl: https://observe.openterms.com/oauth/token
          refreshUrl: https://observe.openterms.com/oauth/token
          scopes:
            read: Read receipts, decisions, policies, keys, workspace.
            write: Create policies, mutate non-admin workspace state.
            admin:write: Rotate keys, activate policies, update workspace settings.

  # --------------------------------------------------------------------------
  # Reusable parameters
  # --------------------------------------------------------------------------
  parameters:
    ReceiptHash:
      name: hash
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-f]{64}$'
        minLength: 64
        maxLength: 64
      description: The canonical hash of the receipt. Exactly 64 lowercase hex characters (SHA-256).
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 128 }
      description: |
        Client-supplied idempotency key. If a request with this key has been
        processed previously and the payload is identical, the stored response
        is returned instead of re-processing. If the payload differs, the
        service returns 409 `IDEMPOTENCY_KEY_CONFLICT`.

  # --------------------------------------------------------------------------
  # Reusable responses
  # --------------------------------------------------------------------------
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            missingField:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Receipt is missing required field 'terms_hash'.
                  details:
                    field: terms_hash
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Request rate limit exceeded.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds until the next request is permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }

  # --------------------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------------------
  schemas:
...
    ReceiptDecision:
      type: string
      enum: [acknowledged, declined]
      description: |
        The receipt-asserted decision (ORS payload field `decision`). Default
        is `acknowledged`. A receipt with `decision: declined` is a refusal
        receipt — proof that an action was declined under a specific policy.
        Distinct from the policy engine's `DecisionOutcome`.
    ReceiptActionType:
      type: string
      enum: [api_call, data_access, purchase, custom, model_training]
      description: |
        The ORS action_type enum. `api_call`, `data_access`, `purchase`, and
        `custom` are defined in ORS v0.1 Section 3a. `model_training` was added
        in ORS v0.2.

    # ----- ORS extension namespaces (under action_context.ors) -----
    ReceiptActionContext:
      type: object
      description: |
        Freeform metadata about the action. MUST have at most 50 top-level
        keys and a serialized size of at most 4096 bytes. MUST NOT contain
        PII. The reserved key `ors` is used for ORS-defined extension
        namespaces (`commitments`, `chain`, `anchor`, `zk_proof`); see ORS
        Section 11. All other keys are issuer-defined.
      properties:
        ors:
          type: object
          description: ORS reserved extension namespace.
          properties:
            commitments: { $ref: '#/components/schemas/OrsCommitments' }
            chain: { $ref: '#/components/schemas/OrsChain' }
            anchor: { $ref: '#/components/schemas/OrsAnchor' }
            zk_proof: { $ref: '#/components/schemas/OrsZkProof' }
          additionalProperties: true
      additionalProperties: true
      maxProperties: 50
...
    Provider:
      type: object
      description: |
        Optional provider identity binding on the signed payload. When present,
        the receipt is bound to a specific provider origin and cannot be
        context-shifted. See ORS Section 3a.1.
      properties:
        origin:
          type: string
          format: uri
          description: |
            Provider HTTPS origin in normalized form: `https://{host}` or
            `https://{host}:{port}`. Host lowercase. Default port (443)
            omitted. No trailing slash. No path.
          pattern: '^https://[a-z0-9.-]+(:[0-9]+)?$'
        provider_id:
          type: string
          description: Provider identifier used by the issuer or provider.
      required: [origin]
      additionalProperties: false
    RequestBinding:
      type: object
      description: |
        Optional binding to a specific provider challenge or request
        commitment. See ORS Section 3a.2. If `binding_method` is
        `provider_nonce`, `provider_nonce` MUST be present. If `request_hash`,
        `request_hash` MUST be present. If `both`, both MUST be present.
      properties:
        provider_nonce: { type: string }
        request_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        expires_at: { type: string, format: date-time }
        binding_method:
          type: string
          enum: [provider_nonce, request_hash, both]
      additionalProperties: false

    # ----- Signed receipt (the wire format) -----
    SignedReceipt:
      type: object
      description: |
        A complete signed ORS receipt as transmitted on the wire. Flat
        structure per ORS Sections 3a–3c. The signed portion (everything
        except `canonical_hash`, `signature`, `key_id`) is canonicalized via
        RFC 8785 plus the ORS null-stripping rule, hashed with SHA-256, and
        signed with Ed25519 over `b"ORSv0.1\x00" + hash_bytes`.

        Accepts both ORS v0.1 and v0.2 receipts. Fields tagged "v0.2" in
        descriptions are optional and only meaningful when `ors_version` is
        `0.2`.
      properties:

        # --- Section 3a: required signed payload ---
        workspace_id:
          type: string
          format: uuid
          description: UUID v4 identifier for the workspace or organization issuing the receipt.
        agent_id:
          type: string
          description: Identifier for the AI agent taking the action.
        action_type:
          $ref: '#/components/schemas/ReceiptActionType'
        terms_url:
          type: string
          format: uri
          description: URL of the terms document being acknowledged. MUST be `http://` or `https://`.
          pattern: '^https?://'
        terms_hash:
          type: string
          description: SHA-256 hash of the terms document content. Lowercase hex, exactly 64 characters. No `sha256:` prefix.
          pattern: '^[0-9a-f]{64}$'
        timestamp:
          type: string
          format: date-time
          description: When the agent requested the receipt. UTC, MUST end with `Z`. Format `YYYY-MM-DDTHH:MM:SS.sssZ`.
        pricing_version:
          type: string
          description: Version identifier for the issuer's pricing schedule.

        # --- Section 3a: optional signed payload ---
        action_context:
          $ref: '#/components/schemas/ReceiptActionContext'
        ors_version:
          type: string
          enum: ['0.1', '0.2']
          description: |
            ORS format version. Optional. If absent, verifiers SHOULD assume
            `0.1`.
        issuer:
          type: string
          format: uri
          description: |
            Issuer identifier for JWKS discovery. MUST be an HTTPS origin (no
            trailing slash, no path). If present, it is signed. For receipts
            issued by this service the value is
            `https://api.observe.openterms.com`.
          pattern: '^https://[a-z0-9.-]+(:[0-9]+)?$'
        provider:
          $ref: '#/components/schemas/Provider'
        decision:
          $ref: '#/components/schemas/ReceiptDecision'
        request_binding:
          $ref: '#/components/schemas/RequestBinding'

        # --- v0.2 optional signed payload ---
        terms_type:
          type: string
          description: |
            v0.2. Kind of policy acknowledged (e.g. "Privacy Policy", "Data
            Processing Agreement", "Terms of Service"). Recommended values
            follow the Open Terms Archive vocabulary.
        terms_service:
          type: string
          description: v0.2. Slug identifying the service whose policy was acknowledged (e.g. `openai`, `aws`, `github`).
        terms_version:
          type: string
          description: v0.2. Version of the policy document at acknowledgment time. Typically a date string `YYYY-MM-DD`. Distinct from `pricing_version`.

        # --- Section 3b: signed envelope fields ---
        receipt_id:
          type: string
          format: uuid
          description: UUID v4 unique identifier for this receipt. Generated by the issuer before signing. Signed.
        amount_charged:
          type: integer
          description: Cost of this receipt in minor currency units. Signed to prevent economic tampering.
        created_at:
          type: string
          format: date-time
          description: Issuer timestamp when the receipt was created. UTC, ends with `Z`. Signed.

        # --- Section 3c: signature metadata (NOT signed) ---
        canonical_hash:
          type: string
          description: |
            SHA-256 of the canonical JSON payload (signed fields only).
            Lowercase hex, 64 characters. NOT included in canonicalization.
          pattern: '^[0-9a-f]{64}$'
        signature:
          type: string
          description: |
            Ed25519 signature, base64url-encoded without padding (RFC 4648
            Section 5). 64 raw signature bytes encode to 86 base64url
            characters. NOT included in canonicalization.
          pattern: '^[A-Za-z0-9_-]+$'
        key_id:
          type: string
          description: |
            Identifier of the signing key used. Maps to the `kid` field of the
            corresponding JWKS entry. NOT included in canonicalization.

      required:
        - workspace_id
        - agent_id
        - action_type
        - terms_url
        - terms_hash
        - timestamp
        - pricing_version
        - receipt_id
        - amount_charged
        - created_at
        - canonical_hash
        - signature
        - key_id
      additionalProperties: true

    # ----- Ingest -----
    IngestResponse:
      type: object
      properties:
        hash:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: Canonical hash of the ingested receipt.
        ingested_at:
          type: string
          format: date-time
        duplicate:
          type: boolean
          description: True if this receipt was already ingested previously (returned unchanged).
        receipt:
          $ref: '#/components/schemas/SignedReceipt'
        decision:
          $ref: '#/components/schemas/Decision'
          description: |
            The engine decision computed at ingest time under the workspace's
            active policy. Optional: omitted only if the engine errored before
            producing a decision (the receipt is still stored and a placeholder
            `deny` decision with an `ENGINE_ERROR` reason is recorded
            server-side for operator alerting).
      required: [hash, ingested_at, duplicate, receipt]

    # ----- Decisions -----
    Decision:
      type: object
      properties:
        decision:
          $ref: '#/components/schemas/DecisionOutcome'
        triggered_rules:
          type: array
          items: { type: string }
          description: Rule IDs that fired during evaluation, in evaluation order.
        reasons:
          type: array
          items: { type: string }
          description: |
            Structured strings explaining why the decision was reached. The
            string `TIMEOUT` appears here when pattern evaluation exceeded its
            time budget (the decision is then `deny`).
        policy_version:
          type: string
        evaluated_at:
          type: string
          format: date-time
      required: [decision, triggered_rules, reasons, policy_version, evaluated_at]
...
    VerificationResult:
      type: object
      description: |
        Mirrors the output of the reference Python verifier
        (`github.com/jstibal/ors-spec/blob/main/verify.py`).
      properties:
        valid: { type: boolean }
        canonical_hash:
          type: [string, "null"]
          description: Recomputed canonical hash. Null if verification failed before hash recomputation.
        key_id:
          type: [string, "null"]
          description: Key ID used for verification. Null if key lookup did not occur.
        receipt_id:
          type: [string, "null"]
          format: uuid
        decision:
          oneOf:
            - $ref: '#/components/schemas/ReceiptDecision'
            - type: "null"
          description: The receipt-asserted decision, if present on the verified receipt.
        jwks_source:
          type: [string, "null"]
          description: The JWKS URL or source identifier used.
        verified_at:
          type: string
          format: date-time
        error:
          type: [string, "null"]
          enum:
            - HASH_MISMATCH
            - KEY_NOT_FOUND
            - UNSUPPORTED_KEY_TYPE
            - INVALID_SIGNATURE
            - INVALID_KEY_LENGTH
            - INVALID_SIGNATURE_LENGTH
            - null
          description: If `valid` is false, the failure category. Null on success.
      required: [valid, canonical_hash, key_id, verified_at, error]

    # ----- Workspace -----
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - HASH_MISMATCH
                - SIGNATURE_INVALID
                - UNKNOWN_ISSUER
                - DUPLICATE
                - IDEMPOTENCY_KEY_CONFLICT
                - POLICY_NOT_FOUND
                - WORKSPACE_NOT_FOUND
                - WEBHOOK_NOT_CONFIGURED
                - RATE_LIMIT_EXCEEDED
                - NOT_FOUND
                - UNAUTHORIZED
                - FORBIDDEN
                - INTERNAL_ERROR
            message: { type: string }
            details:
              oneOf:
                - type: object
                  additionalProperties: true
                - type: "null"
          required: [code, message]
      required: [error]

  # --------------------------------------------------------------------------
  # Examples
  # --------------------------------------------------------------------------
  examples:
    BasicApiCallReceipt:
      summary: Minimal v0.1 api_call receipt (no action_context)
      value:
        workspace_id: 550e8400-e29b-41d4-a716-446655440000
        agent_id: agent-research-bot
        action_type: api_call
        terms_url: https://api.example.com/terms
        terms_hash: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
        timestamp: 2026-05-18T14:22:01.514Z
        pricing_version: 2025-01
        receipt_id: 550e8400-e29b-41d4-a716-446655440010
        amount_charged: 1000
        created_at: 2026-05-18T14:22:01.520Z
        ors_version: '0.1'
        issuer: https://api.observe.openterms.com
        canonical_hash: d486a0a0d298a91b52544aea298693f0ea3584d5f808955d7c09d5f99c16ee32
        signature: c2lnbmF0dXJlX2V4YW1wbGVfYmFzZTY0dXJsX25vcGFkZGluZw
        key_id: key_abc123def456
    DataAccessWithCommitmentsReceipt:
      summary: data_access receipt with ors.commitments under action_context
      value:
        workspace_id: 550e8400-e29b-41d4-a716-446655440000
        agent_id: agent-orchestrator
        action_type: data_access
        terms_url: https://data.example.com/terms
        terms_hash: bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb
        timestamp: 2026-05-18T14:25:00.000Z
        pricing_version: 2025-01
        receipt_id: 550e8400-e29b-41d4-a716-446655440011
        amount_charged: 1000
        created_at: 2026-05-18T14:25:00.120Z
        ors_version: '0.1'
        issuer: https://api.observe.openterms.com
        provider:
          origin: https://data.example.com
          provider_id: provider_data_001
        action_context:
          model: gpt-5.2
          endpoint: /v1/query
          token_count: 1840
          dataset: customer_orders
          ors:
            commitments:
              tool_id: data.query
              args_hash: cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
              pre_state_hash: '1111111111111111111111111111111111111111111111111111111111111111'
              post_state_hash: '2222222222222222222222222222222222222222222222222222222222222222'
        canonical_hash: e5dff2aac4050d92e486761ab21682ea441c3f2e1d8d7d98ac3fd884bbb91d88
        signature: c2lnbmF0dXJlX2V4YW1wbGVfZGF0YV9hY2Nlc3M
        key_id: key_abc123def456
    PolicyClassificationReceipt:
      summary: v0.2 receipt with terms_type / terms_service / terms_version
      value:
        workspace_id: 550e8400-e29b-41d4-a716-446655440000
        agent_id: my-agent-v2
        action_type: api_call
        terms_type: Privacy Policy
        terms_service: openai
        terms_version: '2024-01-15'
        terms_url: https://openai.com/policies/privacy-policy
        terms_hash: cccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccccc
        timestamp: 2026-04-01T12:00:00.000Z
        pricing_version: 2025-01
        receipt_id: 550e8400-e29b-41d4-a716-446655440012
        amount_charged: 0
        created_at: 2026-04-01T12:00:00.050Z
        ors_version: '0.2'
        issuer: https://api.observe.openterms.com
        canonical_hash: 0a1b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f9
        signature: c2lnbmF0dXJlX2V4YW1wbGVfdjAuMg
        key_id: key_abc123def456